+++ NEWS.md
@@ -1,6 +1,7 @@
 ## Version 0.5.1 (2024-09-03)
 
 - implemented `get_news()`
+- implemented `get_ipos()`
 
 ## Version 0.5.0 (2024-08-08)